переименов util и вынес константы в отдельный модуль
