+++ js/data.js
@@ -1,16 +1,5 @@
 import {getRandomArrayElement, getRandomInteger} from './util';
-
-const LIKES_COUNT_MIN = 15;
-const LIKES_COUNT_MAX = 200;
-const COMMENTS_COUNT_MIN = 0;
-const COMMENTS_COUNT_MAX = 200;
-const PHOTO_COUNT = 25;
-
-const DESCRIPTIONS = [
-  'Это невероятно!',
-  'Замечательная работа',
-  'Топ',
-];
+import {LIKES_COUNT_MIN, LIKES_COUNT_MAX, COMMENTS_COUNT_MIN, COMMENTS_COUNT_MAX, PHOTO_COUNT, DESCRIPTIONS} from './constants';
 
 const createPhoto = function (index) {
   return {
